Add sketch for wio terminal
Add multiple methods needed for basic functionallity which will be
outlined here.

setup_wifi(): Add functionallity to connect to a wifi connection. Prints
status text to the serial as the proccess of connecting to wifi is carried out. When a
wifi connection is established this method will print current ip adress
to the serial.

callback(char* topic, byte* payload, unsigned int length): Add
functionallity to revieve message from mqtt broker, then convert the
message from byte* to a string. After that the method
setColorAndPrint(String message) is called with the message as
parameter.

setColorAndPrintMessage(String message): Adds functionallity to fromat
and print a message to the seeed terminal screen.

reconnect(): Add functionallity to connect to MQTT broker. If connection
succeeds a message will be published to specified topic. The specified
topic will then be subscribed to. If connection is not successfull the
terminal will print the state of the MQTT client and then try again
after 5 seconds.

setup(): Adds content to setup method, which will always run when the
terminal is started exactly once. First the screen object is
instanciated, the screen background color is set as well as it's
rotation. Then a serial is instanciated. After that setup_wifi is called
to connect to wifi. Then the MQTT connection is configured. Last the
microphone pin is set to INPUT.

loop(): Add functionallity to firstly see if the terminal is connected
to a wifi network, if not attempt to reconnect. If connected read input
from microphone and then print it to serial. Then wait 0.2 seconds and
then call publishMicValues().

PublishMicValues(): Read microphone value and convert value to char[].
Then publish the char[] to the MQTT broker using the specified topic.