Add functionallity in DataBaseConnection
setUp: Split setup into two parts to make sure client is not out of
scope as it was previously in first setup, also add an attribute to
store client.
addUser: Check if user with given username exists, if not create a
document with given username and password, also make 2 empty lists with
profiles and historicaldata.
verifyLogin: Check if document with given username and password exists,
if so, return true, else return false.