Changed plantCard functionallity
PlantCardController: Change so that the object coresponding to each
card is stored within the card, also add function that takes in
ActionEvent from fxml and calls 2 methods with that event.

plantCard.fxml: Changed what the button calls when pressed, to a method
that calls 2 different methods inside it with the event as parameter.